Only grant permission when `access` returns true
This fixes #635

When you provide an expression inside the `access` option, it will only work as an expression when you prefix it with `@=`. When you forget to do this, it will be just a string.

Previously, the code would check if the access value is trueish instead of strict true. This granted access to all fields that had mistakenly forgotten to prefix with `@=`.