Always pass arguments from expression function
Previously, if I was passing arguments to my resolvers with the
expression function, I would get `null` instead of an instance if the
query has no arguments (e.g. no parens at all in GQL).

However, this breaks resolvers that require an instance argument. This
makes `@arguments` always try to return an instance of an InputObject
whenever it is required.